+++ packages/test/test-end-to-end-tests/package.json
@@ -24,21 +24,21 @@
     "start:tinylicious:test": "tinylicious > tinylicious.log 2>&1",
     "test": "npm run test:realsvc",
     "test:realsvc": "npm run test:realsvc:local && echo 'Skipped: npm run test:realsvc:tinylicious (Issue #6006)'",
-    "test:realsvc:frs": "cross-env fluid__test__driver=r11s fluid__test__r11sEndpointName=frs FLUID_TEST_TIMEOUT=20s npm run test:realsvc:run",
-    "test:realsvc:frs:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:frs",
-    "test:realsvc:local": "cross-env fluid__test__driver=local npm run test:realsvc:run",
-    "test:realsvc:local:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:local",
-    "test:realsvc:odsp": "cross-env fluid__test__driver=odsp FLUID_TEST_TIMEOUT=20s npm run test:realsvc:run",
-    "test:realsvc:odsp:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:odsp",
-    "test:realsvc:r11s": "cross-env fluid__test__driver=r11s FLUID_TEST_TIMEOUT=5000 npm run test:realsvc:run",
+    "test:realsvc:frs": "npm run test:realsvc:run -- --driver=r11s --r11sEndpointName=frs --timeout=20s",
+    "test:realsvc:frs:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:frs --",
+    "test:realsvc:local": "npm run test:realsvc:run -- --driver=local",
+    "test:realsvc:local:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:local --",
+    "test:realsvc:odsp": "npm run test:realsvc:run -- --driver=odsp --timeout=20s",
+    "test:realsvc:odsp:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:odsp --",
+    "test:realsvc:r11s": "npm run test:realsvc:run -- --driver=r11s --timeout=5s",
     "test:realsvc:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc",
     "test:realsvc:routerlicious": "npm run test:realsvc:r11s",
-    "test:realsvc:routerlicious:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:r11s",
+    "test:realsvc:routerlicious:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:r11s --",
     "test:realsvc:run": "mocha dist/test --config src/test/.mocharc.js",
-    "test:realsvc:t9s": "cross-env fluid__test__driver=t9s start-server-and-test start:tinylicious:test 7070 test:realsvc:tinylicious:run",
+    "test:realsvc:t9s": "start-server-and-test start:tinylicious:test 7070 test:realsvc:tinylicious:run",
     "test:realsvc:tinylicious": "npm run test:realsvc:t9s",
     "test:realsvc:tinylicious:report": "cross-env FLUID_TEST_REPORT=1 npm run test:realsvc:t9s",
-    "test:realsvc:tinylicious:run": "npm run test:realsvc:run",
+    "test:realsvc:tinylicious:run": "npm run test:realsvc:run -- --driver=t9s",
     "test:realsvc:verbose": "cross-env FLUID_TEST_VERBOSE=1 npm run test:realsvc",
     "tsfmt": "tsfmt --verify",
     "tsfmt:fix": "tsfmt --replace"
