+++ packages/test/test-end-to-end-tests/src/test/.mocharc.js
@@ -7,8 +7,6 @@
 
 const packageDir = `${__dirname}/../..`;
 
-const common = require("@fluidframework/mocha-test-setup/mocharc-common.js");
-const config = common.getFluidTestMochaConfig(packageDir, [
-    "@fluidframework/test-version-utils"
-]);
+const getFluidTestMochaConfig = require("@fluidframework/test-version-utils/mocharc-common.js");
+const config = getFluidTestMochaConfig(packageDir);
 module.exports = config;
