+++ packages/test/test-version-utils/package.json
@@ -59,6 +59,7 @@
     "@fluidframework/ink": "^0.49.0",
     "@fluidframework/map": "^0.49.0",
     "@fluidframework/matrix": "^0.49.0",
+    "@fluidframework/mocha-test-setup": "^0.49.0",
     "@fluidframework/ordered-collection": "^0.49.0",
     "@fluidframework/protocol-definitions": "^0.1024.0",
     "@fluidframework/register-collection": "^0.49.0",
